Wrapper: re-organize callscript utilities
