+++ packages/aragon-wrapper/src/index.test.js
@@ -3,10 +3,12 @@ import sinon from 'sinon'
 import proxyquire from 'proxyquire'
 import { Subject, empty, of, from } from 'rxjs'
 import { first } from 'rxjs/operators'
-import { encodeCallScript } from './evmscript'
 import { getCacheKey } from './utils'
 import AsyncRequestCache from './utils/AsyncRequestCache'
+import * as callscriptUtils from './utils/callscript'
 import * as forwardingUtils from './utils/forwarding'
+
+const { encodeCallScript } = callscriptUtils
 
 // soliditySha3('app')
 const APP_NAMESPACE_HASH = '0xf1f3eb40f5bc1ad1344716ced8b8a0431d840b5783aea1fd01786bc26f35ac0f'
@@ -26,6 +28,7 @@ test.beforeEach(t => {
     AsyncRequestCache,
     getCacheKey,
     addressesEqual: Object.is,
+    callscript: callscriptUtils,
     forwarding: forwardingUtils,
     makeAddressMapProxy: sinon.fake.returns({}),
     makeProxy: sinon.stub()
